feat(license-signer): introduce license signing tool for evaluation licenses
- Added a new internal tool for generating and signing evaluation license files for the RemoteMedia SDK.
- Implemented keypair generation and license signing functionalities, ensuring secure handling of private keys.
- Included scripts for signing and verifying licenses, with detailed usage instructions in the README.
- Enhanced security measures to prevent accidental exposure of private keys, with clear guidelines for usage.
- Updated documentation to reflect the new tool and its capabilities, providing examples for quick start.

+++ examples/cli/stream-health-demo/src/summary.rs
@@ -157,7 +157,7 @@ impl SessionSummary {
         };
 
         // Estimate samples processed (rough estimate based on event count)
-        let samples_processed = (duration.as_secs() * 16000) as u64; // 16kHz sample rate
+        let samples_processed = duration.as_secs() * 16000; // 16kHz sample rate
 
         Self {
             duration,
@@ -245,6 +245,7 @@ impl SessionSummary {
     }
     
     /// Try to save report, returning a result without failing the session
+    #[allow(dead_code)] // Available for non-failing report saves
     pub fn try_save_report(&self, path: &Path) -> bool {
         match self.save_report(path) {
             Ok(()) => true,
